Added timer text for pause and end level menu

+++ PhaserGame/Assets/Scripts/saveGame.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 [System.Serializable]
 public class saveGame
 {
-    public List<List<int>> levelSave = new List<List<int>>();
+    public List<List<float>> levelSave = new List<List<float>>();
 }
